.release_stage/smart-mail-agent-20250822T203552Z-d64e8c7c/requirements.txt
pydantic>=1.10,<3
python-dotenv>=1.0
reportlab>=3.6,<4
jinja2>=3.1,<4
pytest-cov>=5
PyYAML>=6,<7
pytest-timeout
requests>=2.32
beautifulsoup4>=4.12
Pillow>=10
